requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0
gunicorn>=22.0,<23.0
uvicorn>=0.30,<1.0
psycopg[binary]>=3.2,<4.0

